feat: Add 7z archive creation with streaming support
- Add sevenzip-ffi library with full FFI bindings
- Implement archive creation commands (create_7z_archive, estimate_archive_size)
- Add ExportPanel component for archive creation UI
- Add ActivityProgressPanel for real-time progress tracking
- Implement streaming compression for large archives (>1GB)
- Auto-split archives at 4.7GB for very large files (>8GB)
- Add temp file handling for large archive safety
- Fix FFI bindings to match C library symbols
- Remove old transfer system (replaced by export/archive system)
- Add comprehensive documentation for archive creation feature

Archive Features:
- Multiple compression levels (Store, Fast, Normal, Maximum, Ultra)
- AES-256 encryption support
- Multi-volume split archives
- Progress callbacks with file-level detail
- Memory-efficient streaming for 80GB+ archives

+++ src-tauri/src/ad1/parser.rs
@@ -105,8 +105,8 @@ const ZLIB_CHUNK_ADDR_SIZE: u64 = 0x08;
 pub(crate) struct VerifyParams<'a, F> {
     pub algorithm: HashAlgorithm,
     pub out: &'a mut Vec<VerifyEntry>,
-    pub current: &'a mut usize,
-    pub total: usize,
+    pub current: &'a mut u64,
+    pub total: u64,
     pub progress_callback: &'a mut F,
 }
 
@@ -759,7 +759,7 @@ impl Session {
         params: &mut VerifyParams<'_, F>,
     ) -> Result<(), ContainerError>
     where
-        F: FnMut(usize, usize)
+        F: FnMut(u64, u64)
     {
         let path = join_path(parent_path, &item.name);
         if item.item_type != AD1_FOLDER_SIGNATURE {
@@ -822,12 +822,12 @@ impl Session {
         &mut self,
         item: &Item,
         output_dir: &Path,
-        current: &mut usize,
-        total: usize,
+        current: &mut u64,
+        total: u64,
         progress_callback: &mut F,
     ) -> Result<(), ContainerError>
     where
-        F: FnMut(usize, usize)
+        F: FnMut(u64, u64)
     {
         let item_path = output_dir.join(&item.name);
         if item.item_type == AD1_FOLDER_SIGNATURE {
@@ -930,7 +930,7 @@ mod tests {
     #[test]
     fn test_verify_params_structure() {
         let mut results: Vec<VerifyEntry> = vec![];
-        let mut current: usize = 0;
+        let mut current: u64 = 0;
         
         {
             let params = VerifyParams {
